+++ liboctave/external/ranlib/setall.f
@@ -27,27 +27,27 @@ C                                   INTEGER ISEED1
 C
 C**********************************************************************
 C     .. Parameters ..
-      INTEGER numg
+      INTEGER*4 numg
       PARAMETER (numg=32)
 C     ..
 C     .. Scalar Arguments ..
-      INTEGER iseed1,iseed2
+      INTEGER*4 iseed1,iseed2
       LOGICAL qssd
 C     ..
 C     .. Scalars in Common ..
-      INTEGER a1,a1vw,a1w,a2,a2vw,a2w,m1,m2
+      INTEGER*4 a1,a1vw,a1w,a2,a2vw,a2w,m1,m2
 C     ..
 C     .. Arrays in Common ..
-      INTEGER cg1(numg),cg2(numg),ig1(numg),ig2(numg),lg1(numg),
+      INTEGER*4 cg1(numg),cg2(numg),ig1(numg),ig2(numg),lg1(numg),
      +        lg2(numg)
       LOGICAL qanti(numg)
 C     ..
 C     .. Local Scalars ..
-      INTEGER g,ocgn
+      INTEGER*4 g,ocgn
       LOGICAL qqssd
 C     ..
 C     .. External Functions ..
-      INTEGER mltmod
+      INTEGER*4 mltmod
       LOGICAL qrgnin
       EXTERNAL mltmod,qrgnin
 C     ..
